tzwz/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Babel==3.1.0
Flask-Migrate==4.0.5
WTForms==3.0.1
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
email-validator==2.1.0
Pillow==10.1.0
gunicorn==21.2.0 